+++ unitrait/tests/ui/symbol_prefix_not_string.rs
@@ -1,0 +1,10 @@
+unitrait::unitrait! {
+    #[symbol_prefix = 42]
+    pub trait Foo {
+        pub fn foo() -> u32;
+    }
+
+    macro foo_impl(path = $crate);
+}
+
+fn main() {}